feat(design): Add data/ directory to separate data from component logic.
