+++ src/data/socialLinksData.ts
@@ -1,0 +1,16 @@
+import { SocialMediaType } from '../../common/SocialMedia/SocialMedia';
+
+export const socialLinks: SocialMediaType[] = [
+  {
+    url: 'https://github.com/misael1310',
+    mediaName: 'Github Profile',
+    src: 'https://upload.wikimedia.org/wikipedia/commons/thumb/c/c2/GitHub_Invertocat_Logo.svg/800px-GitHub_Invertocat_Logo.svg.png',
+    alt: 'Github Logo',
+  },
+  {
+    url: 'https://linkedin.com/in/misael-perez-732a4814a/',
+    mediaName: 'Linkedin',
+    src: 'https://upload.wikimedia.org/wikipedia/commons/8/81/LinkedIn_icon.svg',
+    alt: 'Linkedin Logo',
+  },
+];
